% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\docType{data}
\name{mnist_example}
\alias{mnist_example}
\title{MNIST Example Digits}
\format{
A tibble with 7,840 rows and 5 variables:
\describe{
  \item{img_idx}{Image index (1-10)}
  \item{row_idx}{Pixel row position (1-28)}
  \item{col_idx}{Pixel column position (1-28)}
  \item{pixel}{Grayscale pixel intensity (0-255)}
  \item{label}{Digit label (factor): 0-9}
}
}
\source{
\url{http://yann.lecun.com/exdb/mnist/}
}
\usage{
mnist_example
}
\description{
A small sample of 10 handwritten digit images from the MNIST database,
one example of each digit 0-9, in long format.
}
\examples{
data(mnist_example)
table(mnist_example$label)

}
\keyword{datasets}
