% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist.R
\docType{data}
\name{mnist_test}
\alias{mnist_test}
\title{MNIST Test Set (1K subset)}
\format{
A tibble with 784,000 rows and 5 variables:
\describe{
  \item{img_idx}{Image index (1-1000)}
  \item{row_idx}{Pixel row position (1-28)}
  \item{col_idx}{Pixel column position (1-28)}
  \item{pixel}{Grayscale pixel intensity (0-255)}
  \item{label}{Digit label (factor): 0-9}
}
}
\source{
\url{http://yann.lecun.com/exdb/mnist/}
}
\usage{
mnist_test
}
\description{
A 1,000-image subset of the MNIST test database of handwritten
digits, in long format.
}
\examples{
data(mnist_test)
table(mnist_test$label)

}
\keyword{datasets}
